requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.4
typing_extensions==4.15.0
